Consolidated cli libraries.
