+++ container/cli.py
@@ -7,14 +7,22 @@ logger = getLogger(__name__)
 import os
 import sys
 import argparse
+import base64
+import json
 
 import requests.exceptions
 
-from . import core
-from . import config
-from . import exceptions
-from .utils import load_shipit_engine, AVAILABLE_SHIPIT_ENGINES
+import container
 from .conductor import exceptions as conductor_exc
+
+if container.ENV == 'host':
+    from . import core
+    from . import config
+    from . import exceptions
+    from .utils import load_shipit_engine, AVAILABLE_SHIPIT_ENGINES
+elif container.ENV == 'conductor':
+    from .conductor.config import AnsibleContainerConductorConfig
+    from .conductor import core as conductor_core
 
 from logging import config
 LOGGING = {
@@ -205,6 +213,7 @@ class HostCommand(object):
                                help=u'Path to project/context to import.')
 
 
+    @container.host_only
     def __call__(self):
         parser = argparse.ArgumentParser(description=u'Build, orchestrate, run, and '
                                                      u'ship Docker containers with '
@@ -280,3 +289,47 @@ class HostCommand(object):
             sys.exit(1)
 
 host_commandline = HostCommand()
+
+def decode_b64json(encoded_params):
+    return json.loads(base64.decodestring(encoded_params))
+
+@container.conductor_only
+def conductor_commandline():
+    sys.stderr.write('Parsing conductor CLI args.\n')
+    parser = argparse.ArgumentParser(description=u'This should not be invoked '
+                                                 u'except in a container by '
+                                                 u'Ansible Container.')
+    parser.add_argument('command', action='store', help=u'Command to run.',
+                        choices=['build', 'deploy', 'install', 'push', 'run', 'restart',
+                                 'stop'])
+    parser.add_argument('--project-name', action='store', help=u'Project name.', required=True)
+    parser.add_argument('--engine', action='store', help=u'Engine name.', required=True)
+    parser.add_argument('--params', action='store', required=False,
+                        help=u'Encoded parameters for command.')
+    parser.add_argument('--config', action='store', required=True,
+                        help=u'Encoded Ansible Container config.')
+    parser.add_argument('--encoding', action='store', choices=['b64json'],
+                        help=u'Encoding used for parameters.', default='b64json')
+
+    args = parser.parse_args()
+
+    decoding_fn = globals()['decode_%s' % args.encoding]
+    if args.params:
+        params = decoding_fn(args.params)
+    else:
+        params = {}
+
+    if params.get('debug'):
+        LOGGING['loggers']['container']['level'] = 'DEBUG'
+    config.dictConfig(LOGGING)
+
+    containers_config = decoding_fn(args.config)
+    conductor_config = AnsibleContainerConductorConfig(containers_config)
+
+    logger.debug('Starting Ansible Container Conductor: %s', args.command,
+        services=conductor_config.services)
+    getattr(conductor_core, args.command)(args.engine, args.project_name,
+                                          conductor_config.services,
+                                          volume_data=conductor_config.volumes,
+                                          repository_data=conductor_config.registries,
+                                          **params)
